+++ spec/helpers/to_have_vertical_scrollbar.js
@@ -1,14 +1,26 @@
-u = up.util
-
 beforeEach(function() {
   jasmine.addMatchers({
-    toHaveVerticalScrollbar: function(util, customEqualityTesters) {
+    toHaveVerticalScrollbar: function(_util) {
       return {
-        compare: function(element) {
-          let overflow = getComputedStyle(element).overflowY
+        compare: function(viewport) {
+          viewport = up.element.get(viewport)
+
+          // let originalScrollTop = element.scrollTop
+          // element.scrollTop = originalScrollTop === 1 ? 0 : 1
+          // let movedScrollTop = element.scrollTop
+          // element.scrollTop = originalScrollTop
+          //
+          // console.debug("toHaveVerticalScrollbar: element %o, original top %o, moved top %o", element, originalScrollTop, movedScrollTop)
+          //
+          // return {
+          //   pass: movedScrollTop !== originalScrollTop
+          // }
+
+          let overflowElement = viewport === up.viewport.root ? up.viewport.rootOverflowElement() : viewport
+          let overflow = getComputedStyle(overflowElement).overflowY
 
           return {
-            pass: (overflow === 'scroll' || overflow === 'auto') && (element.scrollHeight > element.clientHeight)
+            pass: (overflow === 'scroll' || overflow === 'auto') && (viewport.scrollHeight > viewport.clientHeight)
           }
         }
       }